requirements.txt
Flask==3.0.3
python-pptx==0.6.23
markdown==3.6
